requirements.txt
torch>=2.2
torchvision>=0.17
numpy
pandas
opencv-python
matplotlib
Pillow
